+++ src/main/java/com/javaschool/microecare/usermanagement/repository/TVPPUserRepo.java
@@ -3,7 +3,7 @@ package com.javaschool.microecare.usermanagement.repository;
 import com.javaschool.microecare.usermanagement.dao.TvppUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TVPPUserRepo extends JpaRepository<TvppUser, Integer> {
+public interface TVPPUserRepo extends JpaRepository<TvppUser, Long> {
     TvppUser findByUsername(String username);
 
     boolean existsByUsername(String username);
